backend/requirements.txt
python-dotenv
fastapi 
uvicorn[standard] 
pydantic
pydantic[email]
psycopg